Fix randint dtype issue on windows.
Some more places that need specification of dtype to randint. Problem only manifests on
windows with numpy less that version 2.

+++ brainiak/funcalign/sssrm.py
@@ -351,7 +351,8 @@ class SSSRM(BaseEstimator, ClassifierMixin, TransformerMixin):
         # Initialization:
         self.random_state_ = np.random.RandomState(self.rand_seed)
         random_states = [
-            np.random.RandomState(self.random_state_.randint(2**32))
+            np.random.RandomState(self.random_state_.randint(2**32,
+                                                             dtype=np.int64))
             for i in range(len(data_align))]
         # Set Wi's to a random orthogonal voxels by TRs
         w, _ = srm._init_w_transforms(data_align, self.features, random_states)
